+++ internal/api/oas_router_gen.go
@@ -1461,6 +1461,26 @@ func (s *Server) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 				}
 
+			case 'v': // Prefix: "version"
+
+				if l := len("version"); len(elem) >= l && elem[0:l] == "version" {
+					elem = elem[l:]
+				} else {
+					break
+				}
+
+				if len(elem) == 0 {
+					// Leaf node.
+					switch r.Method {
+					case "GET":
+						s.handleGetVersionRequest([0]string{}, elemIsEscaped, w, r)
+					default:
+						s.notAllowed(w, r, "GET")
+					}
+
+					return
+				}
+
 			case 'w': // Prefix: "workgroups"
 
 				if l := len("workgroups"); len(elem) >= l && elem[0:l] == "workgroups" {
@@ -3290,6 +3310,30 @@ func (s *Server) FindPath(method string, u *url.URL) (r Route, _ bool) {
 
 					}
 
+				}
+
+			case 'v': // Prefix: "version"
+
+				if l := len("version"); len(elem) >= l && elem[0:l] == "version" {
+					elem = elem[l:]
+				} else {
+					break
+				}
+
+				if len(elem) == 0 {
+					// Leaf node.
+					switch method {
+					case "GET":
+						r.name = GetVersionOperation
+						r.summary = ""
+						r.operationID = "getVersion"
+						r.pathPattern = "/version"
+						r.args = args
+						r.count = 0
+						return r, true
+					default:
+						return
+					}
 				}
 
 			case 'w': // Prefix: "workgroups"